rag_pipeline/requirements.txt
langchain
langchain-community
langchain-text-splitters
langchain-openai
langchain-chroma
pypdf
chromadb
python-dotenv
fpdf2
